Now have checkAuthUser, checkAuthTrainer, and checkAuthOwner functions for user verification
